add simple commit check for shell

+++ modules/shell/app/org/corespring/shell/controllers/editor/Item.scala
@@ -1,9 +1,12 @@
 package org.corespring.shell.controllers.editor
 
+import com.mongodb.casbah.Imports._
 import org.bson.types.ObjectId
 import org.corespring.container.client.hooks.Hooks.StatusMessage
 import org.corespring.container.client.hooks.{ItemDraftHooks => ContainerItemDraftHooks, ItemHooks => ContainerItemHooks}
 import org.corespring.mongo.json.services.MongoService
+import org.joda.time.DateTime
+import play.api.Logger
 import play.api.http.Status._
 import play.api.libs.json._
 import play.api.mvc._
@@ -35,6 +38,11 @@ trait ItemHooks extends ContainerItemHooks {
 }
 
 trait ItemDraftHooks extends ContainerItemDraftHooks {
+
+  val logger = Logger("ItemDraftHooks")
+
+  import com.mongodb.casbah.commons.conversions.scala._
+  RegisterJodaTimeConversionHelpers()
 
   def draftItemService: MongoService
 
@@ -80,8 +88,18 @@ trait ItemDraftHooks extends ContainerItemDraftHooks {
     }
   }
 
+  private def addDateModified(itemId:String) = {
+    val builder = DBObject.newBuilder
+    builder += "_id" -> new ObjectId(itemId)
+    builder += "dateModified" -> MongoDBObject("$exists" -> false)
+    itemService.collection.update(builder.result, $set("dateModified" -> DateTime.now), false, false)
+  }
+
   override def create(itemId: String)(implicit h: RequestHeader): R[String] = Future {
     {
+
+      addDateModified(itemId)
+
       for{
         item <- itemService.load(itemId)
         draftId <- draftItemService.create(Json.obj("item" -> item))
@@ -94,13 +112,37 @@ trait ItemDraftHooks extends ContainerItemDraftHooks {
     Right(Json.obj("id" -> draftId))
   }
 
-  override def commit(draftId: String)(implicit h: RequestHeader): R[JsValue] = Future {
+  private def okToCommit(draftId:String) : Boolean = {for{
+      draft <- draftItemService.collection.findOneByID(new ObjectId(draftId), MongoDBObject("item.dateModified" -> 1, "item._id" -> 1))
+      draftItem <- Some(draft.get("item").asInstanceOf[DBObject])
+      draftDateModified <- Some(draftItem.get("dateModified").asInstanceOf[DateTime])
+      itemId <- Some(draftItem.get("_id").asInstanceOf[ObjectId])
+      item <- itemService.collection.findOneByID(itemId, MongoDBObject("dateModified" -> 1))
+      itemDateModified <- Some(item.get("dateModified").asInstanceOf[DateTime])
+    } yield {
+      logger.trace(s"draft date modified: $draftDateModified")
+      logger.trace(s"item date modified: $itemDateModified")
+      val isEqual = draftDateModified.isEqual(itemDateModified.getMillis)
+      val draftIsBefore = draftDateModified.isBefore(itemDateModified.getMillis)
+      logger.trace(s"isEqual: $isEqual, is draft before: $draftIsBefore")
+      isEqual || !draftIsBefore
+    }
+  }.getOrElse(true)
+
+  override def commit(draftId: String, force:Boolean)(implicit h: RequestHeader): R[JsValue] = Future {
     draftItemService.load(draftId).map { draft =>
       val item = (draft \ "item").as[JsObject]
       val itemId = (item \ "_id" \ "$oid").as[String]
-      itemService.save(itemId, item)
-      draftItemService.delete(draftId)
-      Right(Json.obj("itemId" -> itemId, "draftId" -> draftId))
+
+      if(okToCommit(draftId) || force){
+        val updatedItem = item ++ Json.obj("dateModified" -> Json.obj("$date" -> new DateTime().getMillis()))
+        itemService.save(itemId, updatedItem)
+        draftItemService.delete(draftId)
+        Right(Json.obj("itemId" -> itemId, "draftId" -> draftId))
+      } else {
+        Left(BAD_REQUEST, "There has been a new commit since this draft was created")
+      }
+
     }.getOrElse {
       Left((NOT_FOUND, s"Can't find draft by id: $draftId"))
     }
